requirements-dev.txt
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
pyyaml>=6.0,<7.0
black>=24.0,<25.0
flake8>=7.0,<8.0
pylint>=3.0,<4.0
